Move users list to administrator module
