+++ .rubocop.yml
@@ -70,7 +70,7 @@ Rails/ReflectionClassName:
 Rails/SkipsModelValidations:
   Exclude:
     - 'db/migrate/*.rb'
-    - 'app/controllers/users_controller.rb'
+    - 'app/controllers/administrator/users_controller.rb'
 
 Style/Documentation:
   Enabled: false
