+++ app/views/administrator/users/_user.html.erb
@@ -5,12 +5,12 @@
   <td>
     <p>
       <% if user.creation_ip %>
-        <%= t "users.index.summary_html",
+        <%= t ".summary_html",
               :name => link_to(user.display_name, user),
               :ip_address => link_to(user.creation_ip, :ip => user.creation_ip),
               :date => l(user.created_at, :format => :friendly) %>
       <% else %>
-        <%= t "users.index.summary_no_ip_html",
+        <%= t ".summary_no_ip_html",
               :name => link_to(user.display_name, user),
               :date => l(user.created_at, :format => :friendly) %>
       <% end %>
