+++ config/locales/en.yml
@@ -2792,18 +2792,6 @@ en:
       report: Report this User
     go_public:
       flash success: "All your edits are now public, and you are now allowed to edit."
-    index:
-      title: Users
-      heading: Users
-      summary_html: "%{name} created from %{ip_address} on %{date}"
-      summary_no_ip_html: "%{name} created on %{date}"
-      empty: No matching users found
-    page:
-      found_users:
-        one: "%{count} user found"
-        other: "%{count} users found"
-      confirm: Confirm Selected Users
-      hide: Hide Selected Users
     suspended:
       title: Account Suspended
       heading: Account Suspended
@@ -3017,6 +3005,21 @@ en:
       showing_page: "Page %{page}"
       next: "Next"
       previous: "Previous"
+  administrator:
+    users:
+      index:
+        title: Users
+        heading: Users
+        empty: No matching users found
+      page:
+        found_users:
+          one: "%{count} user found"
+          other: "%{count} users found"
+        confirm: Confirm Selected Users
+        hide: Hide Selected Users
+      user:
+        summary_html: "%{name} created from %{ip_address} on %{date}"
+        summary_no_ip_html: "%{name} created on %{date}"
   javascripts:
     close: Close
     share:
